main.cpp
Contains two functions: main and get_list.  main begins by taking in a list of numbers from the user and then proceeds to loop choices for the user to choose until the user exits the program.  get_list is used to create a vector that stores the list that the user inputs.